+++ cudax/examples/stf/linear_algebra/07-cholesky.cu
@@ -93,7 +93,7 @@ public:
 
     for (size_t colb = 0; colb < nt; colb++)
     {
-      int low_rowb = sym_matrix ? colb : 0;
+      size_t low_rowb = sym_matrix ? colb : 0;
       for (size_t rowb = low_rowb; rowb < mt; rowb++)
       {
         T* addr_h = get_block_h(rowb, colb);
@@ -394,23 +394,23 @@ void PDPOTRF(matrix<double>& A)
   assert(A.m == A.n);
   assert(A.mt == A.nt);
 
-  int NBLOCKS = A.mt;
+  size_t NBLOCKS = A.mt;
   assert(A.mb == A.nb);
 
   cuda_safe_call(cudaSetDevice(0));
 
-  for (int K = 0; K < NBLOCKS; K++)
+  for (size_t K = 0; K < NBLOCKS; K++)
   {
     int dev_akk = A.get_preferred_devid(K, K);
     cuda_safe_call(cudaSetDevice(A.get_preferred_devid(K, K)));
     DPOTRF(CUBLAS_FILL_MODE_LOWER, A, K, K);
 
-    for (int row = K + 1; row < NBLOCKS; row++)
+    for (size_t row = K + 1; row < NBLOCKS; row++)
     {
       cuda_safe_call(cudaSetDevice(A.get_preferred_devid(row, K)));
       DTRSM(CUBLAS_SIDE_RIGHT, CUBLAS_FILL_MODE_LOWER, CUBLAS_OP_T, CUBLAS_DIAG_NON_UNIT, 1.0, A, K, K, A, row, K);
 
-      for (int col = K + 1; col < row; col++)
+      for (size_t col = K + 1; col < row; col++)
       {
         cuda_safe_call(cudaSetDevice(A.get_preferred_devid(row, col)));
         DGEMM(CUBLAS_OP_N, CUBLAS_OP_T, -1.0, A, row, K, A, col, K, 1.0, A, row, col);
@@ -551,7 +551,7 @@ void PDGEMM(cublasOperation_t transa,
       //=========================================
       // alpha*A*B does not contribute; scale C
       //=========================================
-      int inner_k = transa == CUBLAS_OP_N ? A.n : A.m;
+      size_t inner_k = transa == CUBLAS_OP_N ? A.n : A.m;
       if (alpha == 0.0 || inner_k == 0)
       {
         DGEMM(transa, transb, alpha, A, 0, 0, B, 0, 0, beta, C, m, n);
