+++ cudax/test/stf/examples/07-cholesky-redux.cu
@@ -554,7 +554,7 @@ void PDGEMM(cublasOperation_t transa,
       //=========================================
       // alpha*A*B does not contribute; scale C
       //=========================================
-      int inner_k = transa == CUBLAS_OP_N ? A.n : A.m;
+      size_t inner_k = transa == CUBLAS_OP_N ? A.n : A.m;
       if (alpha == 0.0 || inner_k == 0)
       {
         DGEMM(transa, transb, alpha, A, 0, 0, B, 0, 0, beta, C, m, n);
